backend/requirements.txt
Django==5.2.12
djangorestframework==3.17.1
django-cors-headers==4.9.0
whitenoise==6.12.0
gunicorn==25.3.0
psycopg2-binary==2.9.11
dj-database-url==3.1.2
